+++ packages/ui/vite.config.ts
@@ -9,10 +9,6 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
-      // "@oraichain/oraidex-evm-sdk": path.resolve(
-      //   __dirname,
-      //   "../sdk/dist/index"
-      // ),
       process: "process/browser",
       buffer: "buffer",
     },
@@ -24,22 +20,40 @@ export default defineConfig({
       fileName: "oraidex-evm-ui",
     },
     rollupOptions: {
-      external: ["react", "react-dom", "buffer", "vite-plugin-node-polyfills/shims/buffer"],
+      external: [
+        "react",
+        "react-dom",
+        "buffer",
+        "vite-plugin-node-polyfills/shims/buffer",
+      ],
       output: {
         globals: {
           react: "React",
           "react-dom": "ReactDOM",
           buffer: "Buffer",
         },
+        assetFileNames: (assetInfo) => {
+          if (assetInfo.name === "style.css") return "oraidex-evm-ui.css";
+          return assetInfo.name;
+        },
       },
     },
+    sourcemap: true,
     commonjsOptions: {
       transformMixedEsModules: true,
     },
+    cssCodeSplit: false,
   },
   optimizeDeps: {},
   server: {
     watch: {},
   },
-  plugins: [react(), dts({ rollupTypes: true }), nodePolyfills()],
+  plugins: [
+    react(),
+    dts({
+      rollupTypes: true,
+      tsconfigPath: "tsconfig.app.json",
+    }) as any,
+    nodePolyfills(),
+  ],
 });